requirements.txt
Pillow>=9.0.0
pytesseract>=0.3.8
